refactored ns
