+++ saq/analysis/__init__.py
@@ -2,4 +2,3 @@ from saq.analysis.analysis import Analysis, ReadOnlyAnalysis
 from saq.analysis.observable import Observable
 from saq.analysis.root import RootAnalysis
 from saq.analysis.tag import Tag
-from saq.analysis.presenter import create_analysis_presenter, AnalysisPresenter, create_observable_presenter, ObservablePresenter
